Adding the ability to show tooltips on less conventional edges
This allows tooltips to be shown on the bottom-left, top-right, etc. NOTE: 'bottom left' is different from 'left bottom' with this implementation. 'bottom left' is the left edge of the bottom side , while 'left bottom' is the bottom edge of the left side. Also note that this hasn't been well tested for tooltips (only for popovers) and the arrow may render incorrectly for them.